Move EngineBuilder into trt_helper.py
